requirements.txt
google-cloud-bigquery>=3.8.0
pandas
streamlit
# BigFrames and Gemini SDKs (if available in your environment)
# bigframes